MAAP-AWS-Anthropic/logger/requirements.txt
fastapi==0.115.6
httpx==0.27.2
pymongo==4.9
python-dotenv==1.0.1
uvicorn
pydantic
motor===3.6.0